Cross Asset Hybrid Models

Cross Asset Hybrid Model Types

DiffFusion.CompositeModel — Type

abstract type CompositeModel <: Model

A CompositeModel represents a collection of (coupled) component models.

CompositeModels are supposed to hold the following elements

alias::String
models::Tuple
state_alias
factor_alias
model_dict::Dict{String,Int}

For concrete types, see SimpleModel and DiagonalModel.

DiffFusion.SimpleModel — Type

struct SimpleModel <: Model
    alias::String
    models::Tuple
    state_alias
    factor_alias
    model_dict::Dict{String,Int}
end

A SimpleModel represents a collection of (coupled) state-independent component models.

It is supposed to be used with a simple_simulation() method.

DiffFusion.simple_model — Function

simple_model(alias::String, models::AbstractVector)

Create a SimpleModel.

DiffFusion.DiagonalModel — Type

DiagonalModel represents a collection of (coupled) component models. For the component models we assume that the models are either state-independent models or diagonal models in the sense that the correlation matrix is state-independent.

For state-independent models we avoid repeated Theta and Sigma calculation.

The model is used as a hybrid model for simulation and payoff evaluation.

DiffFusion.diagonal_model — Function

diagonal_model(alias::String, models::AbstractVector)

Create a DiagonalModel.