Assets and Futures

The following payoffs are used to model FX, Equity, Inflation and Futures instruments.

DiffFusion.Asset — Type

struct Asset <: Leaf
    obs_time::ModelTime
    key::String
end

The price of a tradeable asset S(t) at observation time t.

A tradeable asset is typically an FX rate, equity/index price or spot inflation index.

DiffFusion.ForwardAsset — Type

struct ForwardAsset <: Leaf
    obs_time::ModelTime
    maturity_time::ModelTime
    key::String
end

The forward price E_t[S(T)] of a tradeable asset S at observation time t and with maturity time T. Expectation is calculated in T-forward measure.

DiffFusion.ForwardIndex — Type

struct ForwardIndex <: Leaf
    obs_time::ModelTime
    maturity_time::ModelTime
    key::String
end

Expectation Et^T[ST] of a tradeable asset.

This is used in particular for inflation modelling.

DiffFusion.FutureIndex — Type

struct FutureIndex <: Leaf
    obs_time::ModelTime
    maturity_time::ModelTime
    key::String
end

Risk-neutral expectation Et^T[ST] of a price index.

This is used in particular for Future modelling.

Convexity Adjustment

DiffFusion.AssetConvexityAdjustment — Type

struct AssetConvexityAdjustment <: Leaf
    obs_time::ModelTime
    first_time::ModelTime
    second_time::ModelTime
    pay_time::ModelTime
    key::String
end

Convexity adjustment factor for YoY asset payoffs.

DiffFusion.IndexConvexityAdjustment — Type

struct IndexConvexityAdjustment <: Leaf
    obs_time::ModelTime
    first_time::ModelTime
    second_time::ModelTime
    pay_time::ModelTime
    key::String
end

Convexity adjustment factor for YoY index payoffs.

Asset Option Payoffs

DiffFusion.VanillaAssetOption — Type

struct VanillaAssetOption <: Payoff
    obs_time::ModelTime
    expiry_time::ModelTime
    forward_price::ForwardAsset
    strike_price::Payoff
    call_put::ModelValue
end

The time-t forward price of an option paying [ϕ(F-K)]^+. Forward asset price F is determined at expiry_time.

Option forward price is calculated as expectation in T-forward measure where T corresponds to the expiry time. Conditioning (for time-t price) is on information at obs_time.

Strike price strike_price must be time-t (obs_time) measurable. Otherwise, we look into the future.