Sensitivities

In this section we document sensitivity calculation via Automatic Differentiation (AD).

DiffFusion.model_price — Function

model_price(
    payoffs::AbstractVector,
    path_obj::Path,
    pay_time::Union{ModelTime, Nothing} = nothing,
    discount_curve_key::Union{String,Nothing} = nothing
    )

Calculate model price for a vector of Payoff objects.

DiffFusion.model_price_and_deltas — Function

model_price_and_deltas(
    payoffs::AbstractVector,
    path_obj::Path,
    pay_time::Union{ModelTime, Nothing} = nothing,
    discount_curve_key::Union{String,Nothing} = nothing
    )

Calculate model price and curve sensitivities.

Here, payoffs is a vector of Payoff objects and path_obj is a simulated Path.

pay_time and discount_curve_key control payoff discounting via numeraire calculation.

DiffFusion.model_price_and_vegas — Function

model_price_and_vegas(
    payoffs::AbstractVector,
    model::CompositeModel,
    simulation::Function,
    ts_list::AbstractVector,
    context::Context,
    pay_time::Union{ModelTime, Nothing} = nothing,
    discount_curve_key::Union{String,Nothing} = nothing
    )

Calculate model price and model sensitivities.

Here, payoffs is a vector of Payoff objects, model is a full hybrid model.

simulation is a short cut for a Simulation constructor with the signature simulation(model::Model, ch::CorrelationHolder).

ts_list is a list of TermStructure objects for pricing and context is a Context for valuation.

pay_time and discount_curve_key control payoff discounting via numeraire calculation.

DiffFusion.model_price_and_deltas_vector — Function

model_price_and_deltas_vector(
    payoffs::AbstractVector,
    path_obj::Path,
    pay_time::Union{ModelTime, Nothing} = nothing,
    discount_curve_key::Union{String,Nothing} = nothing,
    ad_module::Module = ForwardDiff,
    )

Calculate model price and curve sensitivities. Sensitivities are calculated as vector together with a vector of labels.

Here, payoffs is a vector of Payoff objects and path_obj is a simulated Path.

pay_time and discount_curve_key control payoff discounting via numeraire calculation.

ad_module can be Zygote or ForwardDiff.

For AMC payoffs we need to update the regression path and trigger a recalibration. For sensitivity calculation, we impose the constraint that regression calibration uses the same paths as valuation.

DiffFusion.model_price_and_vegas_vector — Function

model_price_and_vegas_vector(
    payoffs::AbstractVector,
    model::CompositeModel,
    simulation::Function,
    ts_list::AbstractVector,
    context::Context,
    pay_time::Union{ModelTime, Nothing} = nothing,
    discount_curve_key::Union{String,Nothing} = nothing,
    ad_module::Module = ForwardDiff,
    )

Calculate model price and model sensitivities. Sensitivities are calculated as vector together with a vector of labels.

Here, payoffs is a vector of Payoff objects, model is a full hybrid model.

simulation is a short cut for a Simulation constructor with the signature simulation(model::Model, ch::CorrelationHolder).

ts_list is a list of TermStructure objects for pricing and context is a Context for valuation.

pay_time and discount_curve_key control payoff discounting via numeraire calculation.

ad_module can be Zygote or ForwardDiff.

For AMC payoffs we need to update the regression path and trigger a recalibration. For sensitivity calculation, we impose the constraint that regression calibration uses the same paths as valuation.