Term Structures Functions

In this section we document term structures.

Term Structure Type and Methods

DiffFusion.Termstructure — Type

abstract type Termstructure end

An abstract term structure that provides an alias for identification.

DiffFusion.alias — Method

alias(ts::Termstructure)

Return the term structure's alias.

DiffFusion.TermstructureResultSize — Type

@enum(
    TermstructureResultSize,
    TermstructureVector,
    TermstructureScalar
)

Specify the dimensions/shape of the values modelled by a term structure.

For some term structures (e.g. ParameterTermstructure and VolatilityTermstructure) the result may be either a vector or a scalar.