Analytics

In this section we document methods for instrument pricing, exposure and collateral simulation.

DiffFusion.ScenarioCube — Type

struct ScenarioCube
    X::AbstractArray
    times::AbstractVector
    leg_aliases::AbstractVector
    numeraire_context_key::String
    discount_curve_key::Union{String,Nothing}
end

A ScenarioCube represents the result of MC pricing results of a list of product legs and is calculated for a list of observation times.

Elements are

X - tensor of size (N_1, N_2, N_3) and type ModelValue where

N_1 is number of Monte Carlo paths,

N_2 is number of time steps,

N_3 is number of legs.

times - a vector representing observation times.

leg_aliases - a list of aliases (identifiers) corresponding to each leg

numeraire_context_key - the context_key of the NumeraireEntry; this label should indicate the cash flow currency.

discount_curve_key - a flag specifying whether prices in X are discounted prices (for XVA) or undiscounted prices (for CCR).

DiffFusion.scenarios — Function

scenarios(
    legs::AbstractVector,
    times::AbstractVector,
    path::Path,
    discount_curve_key::Union{String,Nothing};
    with_progress_bar::Bool = true,
    )

Calculate ScenarioCube for a vector of CashFlowLeg objects and a vector of scenario observation times.

DiffFusion.join_scenarios — Function

join_scenarios(cube1::ScenarioCube, cube2::ScenarioCube)

Join two scenario cubes along leg-axis.

join_scenarios(cubes::AbstractVector{ScenarioCube})

Join a list of scenario cubes along leg-axis.

DiffFusion.interpolate_scenarios — Function

interpolate_scenarios(
    t::ModelTime,
    cube::ScenarioCube,
    )

Interpolation scenarios along time axis.

We implement linear interpolation with flat extrapolation.

Other interpolations, e.g., piece-wise flat or Brownian Bridge should be incorporated here.

DiffFusion.concatenate_scenarios — Function

concatenate_scenarios(cubes::AbstractVector{ScenarioCube})

Concatenate a list of scenarios along time axis.

DiffFusion.aggregate — Function

aggregate(
    scens::ScenarioCube,
    average_paths::Bool=true,
    aggregate_legs::Bool=true,
    )

Average paths and aggregate legs in ScenarioCube.

DiffFusion.expected_exposure — Function

expected_exposure(
    scens::ScenarioCube,
    gross_leg::Bool = false,
    average_paths::Bool = true,
    aggregate_legs::Bool = true,
    )

Calculate expected positive exposure (EPE or EE).

DiffFusion.potential_future_exposure — Function

potential_future_exposure(
    scens::ScenarioCube,
    quantile_::ModelValue,
    gross_leg::Bool = false,
    )

Calculate the potential future exposure (PFE).