Examples

In this section we document data and functions to work with the product and model examples.

Product Setup

DiffFusion.Examples.fixed_rate_leg — Function

fixed_rate_leg(
    alias::String,
    effective_time::ModelTime,
    maturity_time::ModelTime,
    coupons_per_year::Int,
    fixed_rate::ModelValue,
    notional::ModelValue,
    discount_curve_key::String,
    fx_key::Union{String, Nothing} = nothing,
    payer_receiver = 1.0,
    )

Create a fixed rate cash flow leg.

DiffFusion.Examples.simple_rate_leg — Function

simple_rate_leg(
    alias::String,
    effective_time::ModelTime,
    maturity_time::ModelTime,
    coupons_per_year::Int,
    forward_curve_key::String,
    fixing_key::Union{String, Nothing},
    spread_rate::Union{ModelValue, Nothing},
    notional::ModelValue,
    discount_curve_key::String,
    fx_key::Union{String, Nothing} = nothing,
    payer_receiver = 1.0,
    )

Create a Libor cash flow leg.

DiffFusion.Examples.compounded_rate_leg — Function

compounded_rate_leg(
    alias::String,
    effective_time::ModelTime,
    maturity_time::ModelTime,
    coupons_per_year::Int,
    forward_curve_key::String,
    fixing_key::Union{String, Nothing},
    spread_rate::Union{ModelValue, Nothing},
    notional::ModelValue,
    discount_curve_key::String,
    fx_key::Union{String, Nothing} = nothing,
    payer_receiver = 1.0,
    )

Create a RFR compounded leg.

DiffFusion.Examples.random_swap — Function

random_swap(example::OrderedDict{String,Any}, type_key::Union{String,Nothing} = nothing)

Sample a random swap.

DiffFusion.Examples.portfolio! — Function

portfolio!(example::OrderedDict{String,Any}, n_swaps::Int = 10)

Create a portfolio of swaps and store it in the dictionary.

DiffFusion.Examples.display_portfolio — Function

Print portfolio in terminal.