Examples (Module)

In this section we document data and functions to work with the product and model examples.

DiffFusion.Examples._csv_path — Constant

const _csv_path = @__DIR__

The path of example CSV files.

DiffFusion.Examples.examples — Constant

A list of example models and instrument specifications.

DiffFusion.Examples.load — Function

load(name::String)

Return a list of dictionaries representing a DiffFusion example.

Example details can be modified by changing the dictionary entries.

DiffFusion.Examples.build — Function

build(dict_list::Vector{OrderedDict{String, Any}})

Return a dictionary of objects and configurations representing a DiffFusion example.

The resulting dictionary is supposed to be queried and amended by methods operating on examples.

DiffFusion.Examples.get_object — Function

Return the first object of a given type from an example dictionary.

Model Setup

DiffFusion.Examples.model — Function

model(example::OrderedDict{String,Any})

Return the hybrid model of an example.

DiffFusion.Examples.correlation_holder — Function

correlation_holder(example::OrderedDict{String,Any})

Return the correlation holder of an example.

DiffFusion.Examples.context — Function

context(example::OrderedDict{String,Any})

Return the context of a given example.

DiffFusion.Examples.term_structures — Function

term_structures(example::OrderedDict{String,Any})

Return a dictionary of term structures for an example.

DiffFusion.Examples.simulation! — Function

simulation!(example::OrderedDict{String,Any})

Return a Monte Carlo simulation for a given example.

If no simulation exists it is created.

DiffFusion.Examples.path! — Function

path!(example::OrderedDict{String,Any})

Return a Monte Carlo path for a given example.

Product Setup

DiffFusion.Examples.fixed_rate_leg — Function

fixed_rate_leg(
    alias::String,
    effective_time::ModelTime,
    maturity_time::ModelTime,
    coupons_per_year::Int,
    fixed_rate::ModelValue,
    notional::ModelValue,
    discount_curve_key::String,
    fx_key::Union{String, Nothing} = nothing,
    payer_receiver = 1.0,
    )

Create a fixed rate cash flow leg.

DiffFusion.Examples.simple_rate_leg — Function

simple_rate_leg(
    alias::String,
    effective_time::ModelTime,
    maturity_time::ModelTime,
    coupons_per_year::Int,
    forward_curve_key::String,
    fixing_key::Union{String, Nothing},
    spread_rate::Union{ModelValue, Nothing},
    notional::ModelValue,
    discount_curve_key::String,
    fx_key::Union{String, Nothing} = nothing,
    payer_receiver = 1.0,
    )

Create a Libor cash flow leg.

DiffFusion.Examples.compounded_rate_leg — Function

compounded_rate_leg(
    alias::String,
    effective_time::ModelTime,
    maturity_time::ModelTime,
    coupons_per_year::Int,
    forward_curve_key::String,
    fixing_key::Union{String, Nothing},
    spread_rate::Union{ModelValue, Nothing},
    notional::ModelValue,
    discount_curve_key::String,
    fx_key::Union{String, Nothing} = nothing,
    payer_receiver = 1.0,
    )

Create a RFR compounded leg.

DiffFusion.Examples.random_swap — Function

random_swap(example::OrderedDict{String,Any}, type_key::Union{String,Nothing} = nothing)

Sample a random swap.

DiffFusion.Examples.portfolio! — Function

portfolio!(example::OrderedDict{String,Any}, n_swaps::Int = 10)

Create a portfolio of swaps and store it in the dictionary.

DiffFusion.Examples.display_portfolio — Function

Print portfolio in terminal.

DiffFusion.Examples.scenarios! — Function

scenarios!(example::OrderedDict{String,Any})

Create the exposure scenarios for the portfolio.